refactor: Move all vendored libs under vendor/
- Move 3rd-party JS code under chrome-ext/src/js/vendor/
- Move 3rd-party CSS files under chrome-ext/src/css/vendor/

This makes it easier to exclude them from linting based on paths
