+++ chrome-ext/tests/test-insert-templates.html
@@ -9,7 +9,7 @@
 
   <link rel="stylesheet" href="./css/ruliweb.css">
   <link rel="stylesheet" href="../templates/css/insert-tables.css">
-  <link rel="stylesheet" href="../src/css/tingle.css">
+  <link rel="stylesheet" href="../src/css/vendor/tingle.css">
   <link rel="stylesheet" href="../src/css/ruliweb-hots.css">
 
   <style>
